Fix Barabasi-Albert bug and increase speed
The previous version of the generator would randomly choose a node for
a potential edge to be made, check against the probability and, if not
successful, simply repeat until a choice was made.

Whilst this works well for smaller graphs the potential for long periods
of time spent polling increases rapidly as the number of nodes does.
This variation uses the WeightedChoice class within JUNG to avoid the
polling and speed up selection of candidate nodes.

Whilst it will have a larger memory footprint it should offer
substantial time improvements for larger graphs.

Also add check to avoid m > m0. Precondition prevents the situation
where the number of requested edges per timestep is greater than can be
fulfilled from the initial seed vertex setup.