Partially format the code base with google-java-format
The Maven plugin https://github.com/coveo/fmt-maven-plugin was used
to do this job. Those files which are still in the process of being
removed or otherwise dealt with for JUNG 3.0 (as described in
https://github.com/jrtom/jung/pull/87) have been left alone to
prevent unnecessary work and merging problems.

The formatting has not yet been tweaked in places where
google-java-format itself cannot automatically do so but which would
be beneficial for readability, like:
1. Removing superfluous empty lines at the start and end of method
   bodies.
2. Adding braces to if, for and while statements.

The formatting has also not been checked to see if javadocs have been
adversely affected.