Remove the 's' from PackageSanityTests
This is admittedly a hacky workaround to get the tests running with
Maven's test runner. Before, they only ever ran in the IDE, so Travis
CI was consequently never running them. Now, they does (or at least
they run through Maven's test runner now!)

The change to ignore `TestUtil` for sanity testing was made in response
to this, for as it turns out, the sanity tests attempt to check that
TestUtil's methods throw NullPointerException, whereas they're not
expected to since they are designed to throw other exceptions instead
for testing purposes.

`testEquals` was also added to each sanity test to hopefully increase
the test coverage a bit.