Fix TreeCollapseDemo (#204)
Previously, when collapsing and expanding the root node of the tree
being viewed, it would throw an unexpected NullPointerException. This
commit fixes things so that the method which was misbehaving,
`TreeUtils#addSubTree`, is able to handle the edge case of populating
an empty tree with a given sub-tree, making the given empty tree
identical to the given sub-tree.